feat: auth types 분리
